// Web/src/generated/model/getCatalogApplicationRuntimeResourceParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * Shared HTTP API contract for the ReleaseHub frontend and backend.
 * OpenAPI spec version: 1.0.0
 */
import type { RuntimeResourceGroupParameter } from './runtimeResourceGroupParameter';
import type { RuntimeResourceKindParameter } from './runtimeResourceKindParameter';
import type { RuntimeResourceNameParameter } from './runtimeResourceNameParameter';
import type { RuntimeResourceNamespaceParameter } from './runtimeResourceNamespaceParameter';
import type { RuntimeResourceVersionParameter } from './runtimeResourceVersionParameter';

export type GetCatalogApplicationRuntimeResourceParams = {
/**
 * @maxLength 253
 */
group?: RuntimeResourceGroupParameter;
/**
 * @minLength 1
 * @maxLength 63
 */
version: RuntimeResourceVersionParameter;
/**
 * @minLength 1
 * @maxLength 63
 */
kind: RuntimeResourceKindParameter;
/**
 * @maxLength 253
 */
namespace?: RuntimeResourceNamespaceParameter;
/**
 * @minLength 1
 * @maxLength 253
 */
name: RuntimeResourceNameParameter;
};